requirements.txt
flask==2.0.1
twilio==7.8.0
pandas==1.3.3
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.24.0
schedule==1.1.0
pytz==2021.3
python-dateutil==2.8.2
requests==2.26.0
python-dotenv==0.19.1
openpyxl==3.0.9
